Ensure FIFO ordering in SynchronousQueue (#234)
By default, SynchronousQueue acts as Stack (LIFO). Only events on the top of a stack will be pop first, causing events to execute in the wrong order (i.e.: PlayerDisconnectEvent executes before PlayerLoginEvent because PDE is on top of the stack).

This commit changes the logic for SynchronousQueue to use an actual Queue instead of a Stack.